refactor: split step utils
