+++ ui/src/features/keys/select-macro/lib/index.ts
@@ -1,2 +1,4 @@
+export * from './convert'
+export * from './normalize'
 export * from './observer'
-export * from './steps'
+export * from './order'
